+++ gems/aws-sdk-configservice/lib/aws-sdk-configservice/types.rb
@@ -3698,6 +3698,8 @@ module Aws::ConfigService
     #
     # * Asia Pacific (Melbourne)
     #
+    # * Canada West (Calgary)
+    #
     # * Europe (Spain)
     #
     # * Europe (Zurich)
@@ -7096,6 +7098,8 @@ module Aws::ConfigService
     #
     #   * Asia Pacific (Melbourne)
     #
+    #   * Canada West (Calgary)
+    #
     #   * Europe (Spain)
     #
     #   * Europe (Zurich)
@@ -7108,9 +7112,9 @@ module Aws::ConfigService
     #
     #    The `AWS::RDS::GlobalCluster` resource type will be recorded in all
     #   supported Config Regions where the configuration recorder is
-    #   enabled, even if `includeGlobalResourceTypes` is not set to `true`.
-    #   The `includeGlobalResourceTypes` option is a bundle which only
-    #   applies to IAM users, groups, roles, and customer managed policies.
+    #   enabled, even if `includeGlobalResourceTypes` is set`false`. The
+    #   `includeGlobalResourceTypes` option is a bundle which only applies
+    #   to IAM users, groups, roles, and customer managed policies.
     #
     #    If you do not want to record `AWS::RDS::GlobalCluster` in all
     #   enabled Regions, use one of the following recording strategies:
@@ -7125,7 +7129,31 @@ module Aws::ConfigService
     #    For more information, see [Selecting Which Resources are
     #   Recorded][1] in the *Config developer guide*.
     #
-    #   <note markdown="1"> Before you set this field to `true`, set the `allSupported` field of
+    #   **includeGlobalResourceTypes and the exclusion recording strategy**
+    #
+    #    The `includeGlobalResourceTypes` field has no impact on the
+    #   `EXCLUSION_BY_RESOURCE_TYPES` recording strategy. This means that
+    #   the global IAM resource types (IAM users, groups, roles, and
+    #   customer managed policies) will not be automatically added as
+    #   exclusions for `exclusionByResourceTypes` when
+    #   `includeGlobalResourceTypes` is set to `false`.
+    #
+    #    The `includeGlobalResourceTypes` field should only be used to
+    #   modify
+    #   the `AllSupported` field, as the default for the `AllSupported`
+    #   field is to record configuration changes for all supported resource
+    #   types excluding the global IAM resource types. To include the global
+    #   IAM resource types when `AllSupported` is set to `true`, make sure
+    #   to set `includeGlobalResourceTypes` to `true`.
+    #
+    #    To exclude the global IAM resource types for the
+    #   `EXCLUSION_BY_RESOURCE_TYPES` recording strategy, you need to
+    #   manually add them to the `resourceTypes` field of
+    #   `exclusionByResourceTypes`.
+    #
+    #   <note markdown="1"> **Required and optional fields**
+    #
+    #    Before you set this field to `true`, set the `allSupported` field of
     #   [RecordingGroup][2] to `true`. Optionally, you can set the `useOnly`
     #   field of [RecordingStrategy][3] to `ALL_SUPPORTED_RESOURCE_TYPES`.
     #
@@ -7288,6 +7316,8 @@ module Aws::ConfigService
     #    * Asia Pacific (Hyderabad)
     #
     #   * Asia Pacific (Melbourne)
+    #
+    #   * Canada West (Calgary)
     #
     #   * Europe (Spain)
     #
@@ -7494,6 +7524,8 @@ module Aws::ConfigService
     #    * Asia Pacific (Hyderabad)
     #
     #   * Asia Pacific (Melbourne)
+    #
+    #   * Canada West (Calgary)
     #
     #   * Europe (Spain)
     #
@@ -8775,8 +8807,7 @@ module Aws::ConfigService
     #   The name or Amazon Resource Name (ARN) of the SSM document to use to
     #   create a conformance pack. If you use the document name, Config
     #   checks only your account and Amazon Web Services Region for the SSM
-    #   document. If you want to use an SSM document from another Region or
-    #   account, you must provide the ARN.
+    #   document.
     #   @return [String]
     #
     # @!attribute [rw] document_version